Move Language model into db/realm.ts
