+++ src/models.ts
@@ -24,7 +24,7 @@ class Fragment {
   public title = "";
   public content = "";
   public snippet!: Snippet; // TODO: Change required to optional?
-  public language?: Language;
+  public language!: Language;
   public createdAt = new Date();
   public updatedAt = new Date();
